requirements.txt
opencv-python~=4.7.0.72
numpy~=1.24.2
streamlit~=1.22.0
pandas
streamlit-drawable-canvas
keras~=2.12.0
tensorflow-macos
